Move getCachedWeatherData function to weather-service.ts
To centralize weather fetching and caching logic within the web service.
